GCC 4.7 SSL fix

+++ src/ssl_support.cc
@@ -355,7 +355,7 @@ ssl_options[] = {
 #endif
 #ifdef SSL_OP_ALL
     {
-        "ALL", SSL_OP_ALL
+        "ALL", (long)SSL_OP_ALL
     },
 #endif
 #ifdef SSL_OP_SINGLE_DH_USE
